# Conflicts:
#	CodeByAI.sketchplugin/Contents/Sketch/library/common.js
#	CodeByAI.sketchplugin/Contents/Sketch/library/i18n/manifest-en.json
#	CodeByAI.sketchplugin/Contents/Sketch/library/i18n/manifest-zh-Hans.json
#	CodeByAI.sketchplugin/Contents/Sketch/library/i18n/manifest-zh-Hant.json
#	CodeByAI.sketchplugin/Contents/Sketch/manifest.json

+++ CodeByAI.sketchplugin/Contents/Sketch/library/generateService.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 sketchDir=$1
-apiUrl="https://service.codebyai.com/api/sketch"
+#apiUrl="https://codebyai.com/api/sketch"
+apiUrl="http://127.0.0.1/api/sketch"
 cd "$sketchDir"
 if [ -f "sketch.zip" ]; then
   rm sketch.zip
@@ -8,13 +9,16 @@ fi
 if [ -d "CodeByAI" ]; then
   rm -rf CodeByAI
 fi
-zip -r sketch.zip  assets/*@2x.png artboard.json *.html.bak
+zip -r sketch.zip  assets/*@2x.png artboard.json bak/*
 curl $apiUrl -X POST -F "sketchZip=@sketch.zip" -o CodeByAI.zip
 unzip CodeByAI.zip
-cp CodeByAI/index.html ./
-rm CodeByAI/index.html
-rm CodeByAI.zip
+if [ $? -eq 0 ]; then
+  cp CodeByAI/index.html ./
+  rm CodeByAI/index.html
+  rm CodeByAI.zip
+fi
 rm sketch.zip
+
 #if [ -f "/usr/local/bin/codebyai-cli" ]; then
 #  /usr/local/bin/codebyai-cli "$sketchDir"
 #fi
